+++ packages/datastore-storage-adapter/CHANGELOG.md
@@ -3,13 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.0.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.0.0...@aws-amplify/datastore-storage-adapter@2.0.1) (2022-11-11)
+## [2.0.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.0.3...@aws-amplify/datastore-storage-adapter@2.0.4) (2022-11-23)
 
 **Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
 
+## [2.0.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.0.2...@aws-amplify/datastore-storage-adapter@2.0.3) (2022-11-19)
 
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
 
+## [2.0.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.0.1...@aws-amplify/datastore-storage-adapter@2.0.2) (2022-11-16)
 
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
+## [2.0.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.0.0...@aws-amplify/datastore-storage-adapter@2.0.1) (2022-11-11)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
 
 # [2.0.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@1.4.5...@aws-amplify/datastore-storage-adapter@2.0.0) (2022-11-09)